# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Begento API",
        "description": "Documentación de la API de Begento",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://begento-test.test",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Obtener mi token de autenticación",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "5fab3ed277fd928880f26f3728c15c33",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear nuevo producto",
                "operationId": "05b2b01e746b59356779b1a7966b18d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "tax_cost",
                                    "manufacturing_cost",
                                    "price",
                                    "currency"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del producto 1"
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "15.5"
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "50.0"
                                    },
                                    "price": {
                                        "required": [
                                            "value"
                                        ],
                                        "properties": {
                                            "value": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 750
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "currency": {
                                        "required": [
                                            "id",
                                            "name",
                                            "symbol",
                                            "exchange_rate"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "US Dollar"
                                            },
                                            "symbol": {
                                                "type": "string",
                                                "example": "$"
                                            },
                                            "exchange_rate": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Laptop X"
                                        },
                                        "price": {
                                            "properties": {
                                                "value": {
                                                    "type": "number",
                                                    "example": 750
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "currency": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "USD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products/id"
                ],
                "summary": "Obtener producto por ID",
                "operationId": "58a864db66b176e2f19a6e2fc072b7ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products/id"
                ],
                "summary": "Actualizar un producto",
                "operationId": "9f350b11a38fe24095d7748904732a3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto existente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "tax_cost",
                                    "manufacturing_cost",
                                    "price",
                                    "currency"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto sustituto"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada del producto"
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "555.0"
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "9999.0"
                                    },
                                    "price": {
                                        "required": [
                                            "value"
                                        ],
                                        "properties": {
                                            "value": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 999999.99
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "currency": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "US Dollar"
                                            },
                                            "symbol": {
                                                "type": "string",
                                                "example": "$"
                                            },
                                            "exchange_rate": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizción exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto actualizado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products/id"
                ],
                "summary": "Borrar producto por ID",
                "operationId": "926f128de2cae5880bcd667913955abb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}/prices": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener lista de precios de un producto",
                "operationId": "1e038105384986347854ecdb0b752af7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios obtenida exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products/id/prices"
                ],
                "summary": "Crear nuevo precio para un producto",
                "operationId": "41eebbb25eefa9281d69ca100c25d481",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "price": {
                                        "properties": {
                                            "value": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 77777.77
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "currency": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "US Dollar"
                                            },
                                            "symbol": {
                                                "type": "string",
                                                "example": "$"
                                            },
                                            "exchange_rate": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Precio nuevo agregado exitosamente"
                    },
                    "400": {
                        "description": "No se pudo agregar el nuevo precio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "login",
            "description": "login"
        },
        {
            "name": "Products",
            "description": "Products"
        },
        {
            "name": "Products/id",
            "description": "Products/id"
        },
        {
            "name": "Products/id/prices",
            "description": "Products/id/prices"
        }
    ]
}